requirements.txt
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
PyYAML>=6.0.1
